requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
discord.py==1.7.3
Flask==2.0.1
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
smmap==4.0.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
yarl==1.6.3
